# [optimize] optimize SimpleModelBuilder's performance by replace strea…