some performance optimization:
- stream/ forEach to for loops
- ConcurrentHashMap's computeIfAbsent to manual no lock get and put
- short circuit for empty data for filterIdSet and filterIdSetOnBuild

before:
```
Benchmark                                  Mode  Cnt        Score        Error  Units
ModelBuilderBenchmark.testBuild           thrpt    3    30854.240 ±  16638.087  ops/s
ModelBuilderBenchmark.testDuplicateMerge  thrpt    3   135970.778 ±  97448.279  ops/s
ModelBuilderBenchmark.testMerge           thrpt    3    68528.907 ±   5372.243  ops/s
ModelBuilderBenchmark.testNullBuild       thrpt    3  1428898.395 ± 149153.447  ops/s
```
after:
```
Benchmark                                  Mode  Cnt        Score        Error  Units
ModelBuilderBenchmark.testBuild           thrpt    3    43123.298 ±   3345.032  ops/s
ModelBuilderBenchmark.testDuplicateMerge  thrpt    3   203827.759 ±   6607.830  ops/s
ModelBuilderBenchmark.testMerge           thrpt    3    93612.534 ±   5639.657  ops/s
ModelBuilderBenchmark.testNullBuild       thrpt    3  1790607.295 ± 554430.074  ops/s
```